requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.11.0
pandas==2.2.1
boto3==1.34.60
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9
apache-airflow-providers-amazon==8.20.0
apache-airflow-providers-postgres==5.10.0
openai==1.14.0
streamlit==1.32.0
plotly==5.20.0